feat(registry): added resource requirements (#70)
- [ ] Changes covered by unit and/or e2e tests;
- [ ] Documentation and examples updated;

## What this PR does / why we need it:
<!--
What code changes are made?
What problem does this PR addresses, or what feature this PR adds?
-->
Add support for specifying resource requests and limits in the Registry
CRD to improve cluster resource utilization and prevent contention in
environments with strict resource quotas.

## Which issue(s) this PR resolves:
<!--
Usage: `Resolves #<issue number>`, or `Resolves <link to the issue>`.
If PR is about `failing-tests`, please post the related tests in a
comment and do not use `Resolves`
-->
Resolves #61

Signed-off-by: Mateusz Urbanek <mateusz.urbanek.98@gmail.com>